Reuse some CLI related code in GUI
Instead of just copypasting functions, attempt to reuse them.
