+++ cli/src/utils.cpp
@@ -23,68 +23,60 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-#include <plorth/gui/context.hpp>
 #include "./utils.hpp"
-#include "../../cli/src/utils.hpp"
 
 namespace plorth
 {
-  namespace gui
+  namespace cli
   {
-    namespace
+    namespace utils
     {
-      class custom_output : public io::output
-      {
-      public:
-        explicit custom_output(Context::text_written_signal signal)
-          : m_signal(signal) {}
-
-        void write(const unistring& text)
-        {
-          m_signal.emit(utils::string_convert<Glib::ustring, unistring>(text));
-        }
-
-      private:
-        const Context::text_written_signal m_signal;
-      };
-    }
-
-    Context::Context()
-      : m_runtime(runtime::make(m_memory_manager))
-      , m_context(context::make(m_runtime))
-    {
-      const auto output = m_runtime->value<custom_output>(
-        m_signal_text_written
-      );
-
-      m_runtime->output() = output;
-      m_runtime->input() = io::input::dummy(m_memory_manager);
-
 #if PLORTH_ENABLE_MODULES
-      plorth::cli::utils::scan_module_path(m_runtime);
-#endif
-    }
-
-    void Context::execute(const Glib::ustring& source_code,
-                          const Glib::ustring& file,
-                          int line)
-    {
-      auto script = m_context->compile(
-        utils::string_convert<unistring, Glib::ustring>(source_code),
-        utils::string_convert<unistring, Glib::ustring>(file),
-        line
-      );
-
-      if (!script || !script->call(m_context))
+      void scan_module_path(const std::shared_ptr<runtime>& rt)
       {
-        auto error = m_context->error();
+#if defined(_WIN32)
+        static const char path_separator = ';';
+#else
+        static const char path_separator = ':';
+#endif
+        auto& module_paths = rt->module_paths();
+        auto begin = std::getenv("PLORTHPATH");
+        auto end = begin;
 
-        m_context->clear_error();
-        if (error)
+        if (end)
         {
-          m_signal_error_thrown.emit(error);
+          for (; *end; ++end)
+          {
+            if (*end != path_separator)
+            {
+              continue;
+            }
+
+            if (end - begin > 0)
+            {
+              module_paths.push_back(utf8_decode(
+                std::string(begin, end - begin)
+              ));
+            }
+            begin = end + 1;
+          }
+
+          if (end - begin > 0)
+          {
+            module_paths.push_back(utf8_decode(
+              std::string(begin, end - begin)
+            ));
+          }
         }
+
+#if defined(PLORTH_RUNTIME_LIBRARY_PATH)
+        if (module_paths.empty())
+        {
+          module_paths.push_back(PLORTH_RUNTIME_LIBRARY_PATH);
+        }
+#endif
       }
+#endif
     }
   }
 }
